Send the aggregate heatmap as float32
Plotly serialises z as a base64 typed array, not decimal text, so the
dtype is the whole cost and rounding buys nothing. 200 x 2,001 float64
made a 5.1 MB payload sent to every browser on the landing page.

float32 halves it: payload 5.15 -> 2.16 MB, figure 8.7 -> 4.8 MB anon.
The values are colour-mapped between 0 and the 99th percentile and
hovered at 4 decimals, so ~7 significant digits is far more than is
visible -- checked, and zero of 400,200 cells land on a different 8-bit
colour step.

The site's other three heatmaps are 19,600 cells at most and two sit on
dead promoter paths, so they are left alone rather than churned.

+++ app/lib/plotting.py
@@ -114,8 +114,15 @@ def fig_aggregate_heatmap(matrix: pd.DataFrame, label: str,
     # else: total_signal (already done)
 
     vmax = float(np.quantile(M.values, 0.99))
+    # float32 on the wire. Plotly serialises z as a base64 TYPED ARRAY
+    # ("dtype":"f8"), not as decimal text, so rounding buys nothing and the
+    # dtype is the whole cost: 200 x 2,001 float64 is a 5.1 MB payload sent
+    # to every browser. The values are colour-mapped between 0 and the 99th
+    # percentile and hovered at 4 decimals, so float32's ~7 significant
+    # digits is far more than is visible. vmax stays float64 -- it is one
+    # number and it sets the scale.
     fig = go.Figure(data=go.Heatmap(
-        z=M.values, x=M.columns.values, y=M.index.tolist(),
+        z=M.values.astype(np.float32), x=M.columns.values, y=M.index.tolist(),
         colorscale=colorscale, zmin=0, zmax=vmax,
         colorbar=dict(title=label, thickness=12),
         hovertemplate="TF: %{y}<br>bp: %{x}<br>signal: %{z:.4f}<extra></extra>",
